+++ packages/core/util/util.js
@@ -15,12 +15,13 @@
  * @exports perfNow
  * @exports mapAndFilterChildren
  * @exports shallowEqual
+ * @exports usePrevious
  */
 import {checkPropTypes as check} from 'prop-types';
 import always from 'ramda/src/always';
 import isType from 'ramda/src/is';
 import unless from 'ramda/src/unless';
-import {Children} from 'react';
+import {Children, useState} from 'react';
 import * as ReactIs from 'react-is';
 
 import Job from './Job';
@@ -361,6 +362,28 @@ const checkPropTypes = (component, props, prevProps) => {
 	}
 };
 
+/**
+ * A custom hook that returns the previous value of a variable.
+ *
+ * @function
+ * @param   {*}    value    The value to track.
+ *
+ * @returns {*}             The value from the previous render.
+ * @memberof core/util
+ * @public
+ */
+const usePrevious = (value) => {
+	const [previousTrackedValue, setPreviousTrackedValue] = useState(value);
+	const [previousValue, setPreviousValue] = useState(value);
+
+	if (value !== previousTrackedValue) {
+		setPreviousTrackedValue(value);
+		setPreviousValue(previousTrackedValue);
+	}
+
+	return previousValue;
+};
+
 export {
 	cap,
 	checkPropTypes,
@@ -375,5 +398,6 @@ export {
 	perfNow,
 	mapAndFilterChildren,
 	setDefaultProps,
-	shallowEqual
+	shallowEqual,
+	usePrevious
 };