+++ CHANGELOG.md
@@ -7,12 +7,18 @@ The following is a curated list of changes in the Enact project, newest changes 
 ### Added
 
 - `core/util` function `checkPropTypes` to check the prop types of a component
+- `ui/resolution` Added support for linear scaling type
+
+### Changed
+
+- `spotlight/Spottable` to reduce redundant work during high-frequency re-renders
 
 ### Fixed
 
 - `spotlight` to focus visible children of elements which do not have `overflow:true`
 - `ui/Marquee.useMarqueeController` not to re-render handlers when any prop is changed
 - `ui/useScroll` to treat the didScroll state update as synchronous
+- `ui/resolution` to display the components accordingly on wide resolutions
 
 ## [5.4.2] - 2026-01-28